requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
celery==5.3.6
redis==5.0.1
requests==2.31.0
zeep==4.2.1  # For SOAP XML integration
prometheus-client==0.19.0
python-jose[cryptography]==3.3.0  # For JWT
python-dotenv==1.0.0
alembic==1.13.1  # For database migrations
aiohttp==3.9.3  # For async HTTP requests 